+++ test/interface/golden/__tests__/sum/SingleSum_spec.ml
@@ -1,0 +1,6 @@
+let () =
+  AesonSpec.goldenDirSpec
+    SingleSum.decodeSingleSum
+    SingleSum.encodeSingleSum
+    "singleSum"
+    "golden/sum/SingleSum";