+++ test/interface/golden/sum/SingleSum.mli
@@ -1,0 +1,6 @@
+type singleSum =
+  | SingleSum
+
+val encodeSingleSum : singleSum -> Js_json.t
+
+val decodeSingleSum : Js_json.t -> (singleSum, string) Belt.Result.t